requirements.txt
alembic
appdirs
attrs
autobahn
Automat
beautifulsoup4
blinker
bs4
cachelib
cachetools
certifi
cffi
charset-normalizer
click
colorama
constantly
contourpy
cryptography
cssselect
curl_cffi
cycler
dhanhq
dnspython
email-validator
fake-useragent
Flask
flask-cors
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Session
Flask-Sessions
Flask-SQLAlchemy
Flask-WTF
fonttools
frozendict
future
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
greenlet
h11
httplib2
hyperlink
idna
incremental
itsdangerous
Jinja2
kiteconnect
kiwisolver
loger
lxml
lxml_html_clean
Mako
MarkupSafe
matplotlib
msgspec
multitasking
numpy
outcome
packaging
parse
peewee
pillow
platformdirs
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
pyOpenSSL
pyotp
pyparsing
pyperclip
pyppeteer
pyquery
PySocks
python-dateutil
python-dotenv
pytz
requests
requests-html
rsa
selenium
service-identity
setuptools
six
smartapi-python
sniffio
sortedcontainers
soupsieve
structures
tabulate
tqdm
trio
trio-websocket
Twisted
txaio
typing_extensions
tzdata
urllib3
w3lib
webdriver-manager
websocket-client
websockets
Werkzeug
wsproto
WTForms
yfinance
zipp
zope.interface
authlib
requests
google-auth-oauthlib
redis
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Session==0.5.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.1.0
pytz==2023.3
Werkzeug==2.3.7
pyotp==2.9.0
razorpay==1.3.0